Added upgrade step to apply our full profile.
This is meant mostly for upgrades from ancient versions that had no
profile yet or had a profile without a metadata.xml.  In that case the
quick installer would complain that the old profile version was
unknown and there was no upgrade.

In a site where I applied this, I noticed the following changes after
running the upgrade:

- An actionicon for CMFPlacefulWorkflow.

- Permission "CMFPlacefulWorkflow: Manage workflow policies" given to
  Manager, Owner and Site Adminstrator.

- The CMFPlacefulWorkflow skin layer was moved below the custom layer.