Fixed undefined name 'portal_url'.
Seen with a code checker, but you get this when you visit the policy mapping form
without a workflow policy id.
Fixed the code up further so you do not get a traceback with a missing or wrong policy id.

Also made sure to call more views with `@@`, which is a tiny bit faster.